moving events to our new events folder
